Feat.MDD123 - Change version number for containers from 3 digits to 2 digits (#42)
# Feat.MDD123 - Change version number for containers from 3 digits to 2 digits
In the past we decided in the main repo to use a 3 digits version number for minor and major changes.
This isn't required and make it more complex. Therefore we decided to change from 3  digit version number to 2 digit version number.

For example: 1.0.1 -> 1.1

### First Digit
The first digit will change if we make major changes for example a new base image. Or any other change where a user need to do any or is a big change.
### Second Digit
The second digit will be changed if we make minor changes. For example we add package to a container or change scripts into the container.